+++ lib/recording_studio_web_search/engine.rb
@@ -103,11 +103,7 @@ module RecordingStudio
 
       initializer "recording_studio_web_search.apply_model_extensions" do
         config.to_prepare do
-          if defined?(RecordingStudioAdmin)
-            require "recording_studio_web_search/admin"
-            Admin.register!
-          end
-          RunPage.prepare!
+          HostBoot.prepare!
           next unless defined?(ActiveRecord::Base)
 
           ActiveRecord::Base.descendants.each do |model|